Added support for zooming and downloading media
