+++ lib/tweet/_video.dart
@@ -1,0 +1,97 @@
+import 'package:dart_twitter_api/twitter_api.dart';
+import 'package:flutter/material.dart';
+import 'package:video_player/video_player.dart';
+
+class TweetVideo extends StatefulWidget {
+  final bool loop;
+  final Media media;
+
+  const TweetVideo({Key? key, required this.loop, required this.media}) : super(key: key);
+
+  @override
+  State<StatefulWidget> createState() => _TweetVideoState();
+}
+
+class _TweetVideoState extends State<TweetVideo> {
+  late VideoPlayerController _controller;
+
+  @override
+  void initState() {
+    super.initState();
+
+    _controller = VideoPlayerController.network(widget.media.videoInfo!.variants![0].url!);
+    _controller.setLooping(widget.loop);
+    _controller.initialize().then((_) {
+      // Ensure the first frame is shown after the video is initialized, even before the play button has been pressed.
+      setState(() {});
+    });
+
+    _controller.addListener(() {
+      setState(() {});
+    });
+  }
+
+  Widget _createControls() {
+    return Stack(
+      children: <Widget>[
+        AnimatedSwitcher(
+          duration: Duration(milliseconds: 50),
+          reverseDuration: Duration(milliseconds: 200),
+          child: _controller.value.isPlaying
+              ? SizedBox.shrink()
+              : Container(
+            color: Colors.black26,
+            child: Center(
+              child: Icon(
+                Icons.play_arrow,
+                color: Colors.white,
+                size: 100.0,
+              ),
+            ),
+          ),
+        ),
+        GestureDetector(
+          onTap: () {
+            _controller.value.isPlaying
+                ? _controller.pause()
+                : _controller.play();
+          },
+        ),
+      ],
+    );
+
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    double aspectRatio = widget.media.videoInfo?.aspectRatio == null
+        ?  _controller.value.aspectRatio
+        : widget.media.videoInfo!.aspectRatio![0] / widget.media.videoInfo!.aspectRatio![1];
+
+    return GestureDetector(
+      onTap: () {
+        setState(() {
+          _controller.value.isPlaying
+              ? _controller.pause()
+              : _controller.play();
+        });
+      },
+      child: AspectRatio(
+        aspectRatio: aspectRatio,
+        child: Stack(
+          alignment: Alignment.bottomCenter,
+          children: [
+            VideoPlayer(_controller),
+            _createControls(),
+          ],
+        ),
+      ),
+    );
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+    _controller.dispose();
+  }
+}
